@@ -13,8 +13,8 @@
   },
   "dependencies": {
     "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.6.0",
-    "@fontsource/lato": "^4.5.1",
+    "@emotion/styled": "^11.8.1",
+    "@fontsource/lato": "^4.5.3",
     "@loadable/component": "^5.15.2",
     "@mui/material": "^5.5.1",
     "@perses-dev/components": "^0.1.0",
@@ -25,7 +25,7 @@
     "@perses-dev/prometheus-plugin": "^0.1.0",
     "immer": "^9.0.12",
     "lodash-es": "^4.17.21",
-    "mdi-material-ui": "^7.0.0",
+    "mdi-material-ui": "^7.2.0",
     "notistack": "^2.0.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -38,14 +38,15 @@
     "@svgr/webpack": "^6.2.1",
     "@types/loadable__component": "^5.13.4",
     "@types/webpack-bundle-analyzer": "^4.4.1",
-    "css-loader": "^6.6.0",
+    "css-loader": "^6.7.1",
     "dotenv-defaults": "^5.0.0",
-    "fork-ts-checker-webpack-plugin": "^6.5.0",
+    "eslint-webpack-plugin": "^3.1.1",
+    "fork-ts-checker-webpack-plugin": "^7.2.1",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.6",
+    "ts-loader": "^9.2.8",
     "ts-node": "^10.7.0",
-    "webpack": "^5.68.0",
+    "webpack": "^5.70.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4",
